change color handling module
