+++ src/visGeometry/VisAgent.ts
@@ -49,7 +49,10 @@ export default class VisAgent {
     public resetMesh(): void {
         this.followed = false;
         this.highlighted = false;
-        this.setColor(new Color(VisAgent.UNASSIGNED_MESH_COLOR), 0);
+        this.setColor({
+            color: new Color(VisAgent.UNASSIGNED_MESH_COLOR),
+            colorId: 0,
+        });
         this.agentData = {
             x: 0,
             y: 0,
@@ -65,9 +68,9 @@ export default class VisAgent {
         };
     }
 
-    public setColor(color: Color, colorIndex: number): void {
-        this.color = color;
-        this.colorIndex = colorIndex;
+    public setColor(colorInfo: { color: Color; colorId: number }): void {
+        this.color = colorInfo.color;
+        this.colorIndex = colorInfo.colorId;
     }
 
     public setHidden(hidden: boolean): void {
